Added code to verify the timestamp and seperated main , server thread

+++ src/Message.java
@@ -1,5 +1,6 @@
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * 
@@ -13,10 +14,10 @@ enum MessageType {
 	REQUEST_KEY,RESPONSE_KEY,RESPONSE_AND_REQUEST_KEY,TERMINATION_REQUEST, TERMINATION_RESPONSE;
 }
 public class Message implements Serializable{
-	long timeStamp;
+	AtomicInteger timeStamp;
 	MessageType messageType;
 	Host nodeInfo;
-	public Message(long timeStamp, MessageType messageType, Host nodeInfo)
+	public Message(AtomicInteger timeStamp, MessageType messageType, Host nodeInfo)
 	{
 		this.timeStamp = timeStamp;
 		this.messageType = messageType;
